The code I had created to pull the client-side proxy of botania without
crashing server-side was still crashing server-side.  For some reason, I
thought everything would be fine so long as I didn't instantiate the proxy
but of course, even pulling the class will try to load the dependencies,
which aren't present.  Now we're only pulling the class in our own client
proxy.